move `isRemixOutletElement` to `project-file-utils`
