+++ editor/src/core/model/project-file-utils.ts
@@ -183,6 +183,19 @@ export function isRemixOutletAgainstImports(element: JSXElementChild, imports: I
   )
 }
 
+export function isRemixOutletElement(
+  element: JSXElementChild,
+  filePath: string,
+  projectContents: ProjectContentTreeRoot,
+): boolean {
+  const file = getProjectFileByFilePath(projectContents, filePath)
+  if (file != null && isTextFile(file) && isParseSuccess(file.fileContents.parsed)) {
+    return isRemixOutletAgainstImports(element, file.fileContents.parsed.imports)
+  } else {
+    return false
+  }
+}
+
 export function isEllipseAgainstImports(jsxElementName: JSXElementName, imports: Imports): boolean {
   return isGivenUtopiaAPIElementFromName(jsxElementName, imports, 'Ellipse')
 }
